+++ src/contracts/VotingAlgMajority.sol
@@ -19,7 +19,7 @@ import "./VotingAlgInterface.sol";
 */
 contract VotingAlgMajority is VotingAlgInterface {
 
-    function assess(uint32 numParticipants, uint32 numVotedFor, uint32 /* numVotedAgainst */) external pure returns (bool) {
+    function assess(uint64 numParticipants, uint64 numVotedFor, uint64 /* numVotedAgainst */) external pure returns (bool) {
         return (numVotedFor * 2 > numParticipants);
     }
 }
